🎨: recipe-stepコンポーネントに修正
