+++ src/features/recipes/index.ts
@@ -1,3 +1,3 @@
 // components:
 export * from "./components/recipe-favorite-button";
-export * from "./components/recipe-instruction";
+export * from "./components/recipe-step";
